Add default value for service description variable
When the definition of the variable `app_description` was missing,
ansible would complain and abort the execution of the playbook.

Although describing the service is surely recommended, it should not
be mandatory. Therefore a more or less reasonable default has been
defined which is used in the upstart and the systemd service
definitions.

Signed-off-by: Christian Harke <christian.harke@bluecare.ch>